Fix code climate issues + add system utility tests
-> Split large files into smaller submodules, and large functions into smaller functions

-> Remove unused and unneeded modules, classes and function parameters

-> Remove some optional instance parameters from the constructors of classes to reduce bloat (VideoWriter, MusicVideoGenerator). These can instead be set via dot notation after the instance is created.

-> Add dedicated "transformation" module for applying sizing and effects to segments

-> Split events module into several submodules

-> Move video effect application functions under their respective effect class definitions

+ Use the same, single subtitles track for both music videos and previews

+ Remove preview mode to support only audiovisual previews

+ Add sample_with_filters function to SourceSampler

+ Store events used to generate the music video on the music video instance

+ Rename detect module to detection

+ Add test for `--use-original-audio` option

+ Add extra speed multiply test cases for events
