+++ README.md
@@ -195,7 +195,8 @@ generator = MusicVideoGenerator("Pogo - Forget.mp3")
 beats = generator.audio.beats()
 beats.speed_multiply(1/2)
 
-generator.preview_events(beats, "forget-preview.mkv")
+preview = generator.preview_from_events(beats, "forget-preview.mkv")
+preview.write_to_video_file("preview.mkv")
 ```
 
 ### Create a music video
@@ -261,7 +262,7 @@ music_video.segments[1].ipython_display(t=.5, width=400)
 
 ### Subtitles
 
-The music videos generated by `create` include subtitles which display segment numbers, locations, and intervals. Likewise the previews generated by `preview` include subtitles which display detected event types, numbers, and locations.
+The videos generated by `create` and `preview` include a subtitle track which display segment types, numbers, and locations.
 
 ### Text detection
