+++ mugen/constants.py
@@ -5,9 +5,9 @@ from typing import Tuple, Union
 PLATFORM = platform.system()
 
 """
-SEC.MIL or 'SEC.MIL'
-(MIN, SEC.MIL) or 'MIN:SEC.MIL'
-(HRS, MIN, SEC.MIL) or 'HRS:MIN:SEC.MIL'
+SECONDS.MILLISECONDS or 'SECONDS.MILLISECONDS'
+(MINUTES, SECONDS.MILLISECONDS) or 'MINUTES:SECONDS.MILLISECONDS'
+(HOURS, MINUTES, SECONDS.MILLISECONDS) or 'HOURS:MINUTES:SECONDS.MILLISECONDS'
 """
 TIME_FORMAT = Union[float, Tuple[int, float], Tuple[int, int, float], str]
