+++ mugen/video/detection.py
