+++ mugen/video/effects.py
@@ -1,7 +1,7 @@
-from typing import List, Optional
+from abc import ABC, abstractmethod
+from typing import List
 
 from mugen.constants import Color
-from mugen.lists import MugenList
 from mugen.utilities import conversion
 from mugen.utilities.conversion import convert_color_to_hex_code
 from mugen.video.events import VideoEvent
@@ -9,62 +9,19 @@ from mugen.video.events import VideoEvent
 ONE_SECOND = 1
 
 
-class VideoEffect(VideoEvent):
+class VideoEffect(VideoEvent, ABC):
     """
     A visual effect in a video
     """
 
     pass
 
-
-class VideoEffectList(MugenList):
-    """
-    Wrapper for a list of Effects, providing convenience methods
-    """
-
-    def __init__(self, effects: Optional[List[VideoEffect]] = None):
-        super().__init__(effects)
-
-    def add_crossfade(self, duration: float = ONE_SECOND):
+    @abstractmethod
+    def apply(self):
         """
-        Adds a crossfade effect to the beginning of the video segment.
-
-        Parameters
-        ----------
-        duration
-            duration of the effect
+        Applies the effect to a segment
         """
-        self.append(CrossFade(duration))
-
-    def add_fadein(self, duration: float = ONE_SECOND, color: str = Color.BLACK):
-        """
-        Adds a fade-in effect to the video segment.
-
-        Parameters
-        ----------
-        duration
-            duration of the effect
-
-        color
-            hex code of the color to fade in from, i.e. #000000 for black.
-            OR the special inputs 'black' and 'white'.
-        """
-        self.append(FadeIn(duration, color))
-
-    def add_fadeout(self, duration: float = ONE_SECOND, color: str = Color.BLACK):
-        """
-        Adds a fade-out effect to the video segment.
-
-        Parameters
-        ----------
-        duration
-            duration of the effect
-
-        color
-            hex code of the color to fade out to, i.e. #000000 for black.
-            OR the special inputs 'black' and 'white'.
-        """
-        self.append(FadeOut(duration, color))
+        pass
 
 
 class Fade(VideoEffect):
@@ -81,15 +38,86 @@ class Fade(VideoEffect):
 
 
 class FadeIn(Fade):
-    def __init__(self, duration: float, color: str, **kwargs):
+    """
+    Adds a fade-in effect to the video segment
+    """
+
+    def __init__(
+        self, duration: float = ONE_SECOND, color: str = Color.BLACK, **kwargs
+    ):
+        """
+        Parameters
+        ----------
+        duration
+            duration of the effect
+
+        color
+            hex code of the color to fade in from, i.e. #000000 for black.
+            OR the special inputs 'black' and 'white'.
+        """
         super().__init__(duration=duration, color=color, **kwargs)
+
+    def apply(self, segment):
+        segment = segment.fadein(self.duration, self.rgb_color)
+        if segment.audio:
+            segment.audio = segment.audio.audio_fadein(self.duration)
+
+        return segment
 
 
 class FadeOut(Fade):
-    def __init__(self, duration: float, color: str, **kwargs):
+    """
+    Adds a fade-out effect to the video segment
+    """
+
+    def __init__(
+        self, duration: float = ONE_SECOND, color: str = Color.BLACK, **kwargs
+    ):
+        """
+        Parameters
+        ----------
+        duration
+            duration of the effect
+
+        color
+            hex code of the color to fade out to, i.e. #000000 for black.
+            OR the special inputs 'black' and 'white'.
+        """
         super().__init__(duration=duration, color=color, **kwargs)
 
+    def apply(self, segment):
+        segment = segment.fadeout(self.duration, self.rgb_color)
+        if segment.audio:
+            segment.audio = segment.audio.audio_fadeout(self.duration)
 
-class CrossFade(VideoEffect):
-    def __init__(self, duration: float, **kwargs):
+        return segment
+
+
+class Crossfade(VideoEffect):
+    """
+    Adds a crossfade effect to the beginning of the video segment
+    """
+
+    def __init__(self, duration: float = ONE_SECOND, **kwargs):
+        """
+        Parameters
+        ----------
+        duration
+            duration of the effect
+        """
         super().__init__(duration=duration, **kwargs)
+
+    def apply(self, segment, previous_segment):
+        segment = segment.set_start(previous_segment.end - self.duration)
+        segment = segment.crossfadein(self.duration)
+        if segment.audio:
+            segment = segment.set_audio(segment.audio.audio_fadein(self.duration))
+
+        return segment
+
+    def buffer(self, segment):
+        buffer = segment.trailing_buffer(self.duration)
+        if buffer.audio:
+            buffer = buffer.set_audio(buffer.audio.audio_fadeout(self.duration))
+
+        return buffer
