+++ mugen/video/events.py
@@ -1,4 +1,5 @@
-from mugen.events import Event
+from mugen.constants import TIME_FORMAT
+from mugen.events.Event import Event
 
 
 class VideoEvent(Event):
@@ -7,3 +8,12 @@ class VideoEvent(Event):
     """
 
     pass
+
+
+class Cut(VideoEvent):
+    """
+    A cut from one piece of sources to another in a video
+    """
+
+    def __init__(self, location: TIME_FORMAT, *args, **kwargs):
+        super().__init__(location, *args, **kwargs)
