+++ tests/end_to_end/cli/test_create.py
@@ -67,23 +67,13 @@ def test_create__creates_music_video_successfully(tmp_path):
     assert music_video_segment.video_stream["height"] == 600
     assert music_video_segment.video_stream["codec_name"] == "hevc"
 
-    # Check subtitle tracks
-    assert len(music_video_segment.streams) == 5
+    # Check video, audio and subtitle tracks
     assert len(music_video_segment.video_streams) == 1
     assert len(music_video_segment.audio_streams) == 1
-    assert len(music_video_segment.subtitle_streams) == 3
-    assert music_video_segment.subtitle_streams[0]["tags"]["title"] == "segment_numbers"
-    assert (
-        music_video_segment.subtitle_streams[1]["tags"]["title"]
-        == "segment_locations_seconds"
-    )
-    assert (
-        music_video_segment.subtitle_streams[2]["tags"]["title"]
-        == "segment_durations_seconds"
-    )
+    assert len(music_video_segment.subtitle_streams) == 1
+    assert len(music_video_segment.streams) == 3
+    assert music_video_segment.subtitle_streams[0]["tags"]["title"] == "events"
     assert len(music_video_segment.get_subtitle_stream_content(0)) > 0
-    assert len(music_video_segment.get_subtitle_stream_content(1)) > 0
-    assert len(music_video_segment.get_subtitle_stream_content(2)) > 0
 
 
 def test_create__works_with_files_with_special_characters(tmp_path):
@@ -118,3 +108,54 @@ def test_create__works_with_files_with_special_characters(tmp_path):
     # Check that output files exist
     assert os.path.isfile(music_video_path)
     assert os.path.isfile(music_video_save_file_path)
+
+
+def test_create__preserves_original_audio_when_option_is_passed(tmp_path):
+    audio_path = TWO_BEATS_AUDIO_PATH
+    try:
+        subprocess.run(
+            [
+                "mugen",
+                "--output-directory",
+                tmp_path,
+                "create",
+                "--audio-source",
+                audio_path,
+                "--video-sources",
+                TRACKING_SHOT_VIDEO_PATH,
+                "--use-original-audio",
+            ],
+            check=True,
+            timeout=180,
+            capture_output=True,
+            text=True,
+        )
+    except CalledProcessError as error:
+        print(error.stdout)
+        print(error.stderr)
+        raise error
+
+    music_video_path_base = os.path.join(tmp_path, "music_video_0", "music_video_0")
+    music_video_path = f"{music_video_path_base}.mkv"
+    music_video_save_file_path = f"{music_video_path_base}.pickle"
+
+    # Check that output files exist
+    assert os.path.isfile(music_video_path)
+    assert os.path.isfile(music_video_save_file_path)
+
+    music_video_segment = VideoSegment(music_video_path)
+    loaded_music_video = MusicVideo.load(music_video_save_file_path)
+
+    # Check duration
+    assert len(loaded_music_video.segments) == 3
+    assert music_video_segment.duration == pytest.approx(
+        loaded_music_video.duration, 0.1
+    )
+
+    # Check video, audio and subtitle tracks
+    assert len(music_video_segment.video_streams) == 1
+    assert len(music_video_segment.audio_streams) == 1
+    assert len(music_video_segment.subtitle_streams) == 1
+    assert len(music_video_segment.streams) == 3
+    assert music_video_segment.subtitle_streams[0]["tags"]["title"] == "events"
+    assert len(music_video_segment.get_subtitle_stream_content(0)) > 0
